+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 * touch - added an empty (disable) classifier, exposed more profile
   values dynamically, option to hook profile picking on unknown device.
 
+* new tool - text to speech
+
 minor:
 
 * added dracula/gruvbox/gruvbox-light colorschemes